+++ meta/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the Orbital Trading project will be documented in this file by the user, only. Gemini is never to modify or provide updates to this file, and must always treat this file as read-only. The user will always manually update this file.
 
+# [36.39] - 2026-3-2
+-  added intro cinematic
+-  added skip functionality
+-  added web view method
+-  added xcode view method
+
 # [36.35] - 2026-2-27
 -  added system states v2
 -  added economic weather interface